MetricMatrix/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.8
django-auto-logout==0.5.0
django-axes==5.27.0
django-guardian==2.4.0
django-ipware==4.0.0
django-oauth-toolkit==1.5.0
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-guardian==0.3.0
drf-api-logger==1.0.8
idna==3.3
jwcrypto==1.0
mod-wsgi==4.9.0
oauthlib==3.1.1
packaging==21.3
pycparser==2.20
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
webencodings==0.5.1
wrapt==1.13.3
